requirements.txt
alembic
APScheduler
bcrypt
cloudinary
email_validator
emails
fastapi
fastapi-events
itsdangerous
Jinja2
phonenumbers
psycopg2-binary
pydantic
pydantic-settings
PyJWT
python-multipart
requests
SQLAlchemy
starlette
typing_extensions
uvicorn